add triangle

+++ Computer Graphics/Triangle 3D/src/Shape.cpp
@@ -1,5 +1,6 @@
 #include "Shape.h"
 
-int Shape::m_Height = Globals::Height * .8f;
+int Shape::m_Height = static_cast<int>(Globals::Height * .8f);
 int Shape::m_Width = m_Height;
 int Shape::m_Position = 0;
+
